requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
dj-rest-auth==6.0.0
Django==4.2.16
django-allauth==65.0.1
django-axes==6.5.1
django-cors-headers==4.4.0
django-environ==0.11.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
psycopg2-binary==2.9.9
PyJWT==2.9.0
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.1
